Visual Studio builds: Require Visual Studio 2017
cairomm 1.16 requires libsigc++-3.0, which requires C++-14 features that
are supported with Visual Studio 2017 or later, so update the projects
and README files accordingly.  Also update the DLL version as the ABI
version was updated for 1.16.

Also, use the property sheets to define the version of libsigc++ so that
it is easier to update it in case we need to in the future.

Bug 100210