build: Add NMake Makefiles for MSVC builds
This adds a set of NMake Makefiles to build cairomm, and its example
programs and its test programs (the test programs are not buildable
in their current state).  Visual Studio 2017 is still the minimum
required version of Visual Studio required, with verion 15.8.0 or
later recommended.

The NMake Makefiles will consume cairomm/filelist.am, so that additions
and removal of sources will only need to be done there, which will
reduce the likelyhood of the MSVC build files getting out-of-date.

Dist only the NMake Makefiles, and also dist the gendef,
cairommconfig.h and cairomm.rc(.in) as we did before.  The Visual
Studio projects will be dropped in the next commit.